[java-utils-scanner] patched PackageScanner & updated dependencies
