+++ java-utils-scanner/src/main/java/net/runeduniverse/lib/utils/scanner/ScanOrder.java
@@ -16,5 +16,7 @@
 package net.runeduniverse.lib.utils.scanner;
 
 public enum ScanOrder {
+
 	FIRST, LAST, ALL
+
 }
