+++ java-utils-scanner/src/main/java/net/runeduniverse/lib/utils/scanner/api/IFieldScanner.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.runeduniverse.lib.utils.scanner;
+package net.runeduniverse.lib.utils.scanner.api;
 
 import java.lang.reflect.Field;
 
@@ -21,5 +21,7 @@ import net.runeduniverse.lib.utils.scanner.pattern.FieldPattern;
 import net.runeduniverse.lib.utils.scanner.pattern.TypePattern;
 
 public interface IFieldScanner<F extends FieldPattern> {
+
 	void scan(Field field, Class<?> type, TypePattern<F, ?> pattern) throws Exception;
+
 }
