+++ java-utils-scanner/src/main/java/net/runeduniverse/lib/utils/scanner/api/IMethodScanner.java
@@ -13,8 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.runeduniverse.lib.utils.scanner;
+package net.runeduniverse.lib.utils.scanner.api;
 
-public interface ITypeScanner {
-	void scan(Class<?> type, ClassLoader loader, String pkg) throws Exception;
+import java.lang.reflect.Method;
+
+import net.runeduniverse.lib.utils.scanner.pattern.MethodPattern;
+import net.runeduniverse.lib.utils.scanner.pattern.TypePattern;
+
+public interface IMethodScanner<M extends MethodPattern> {
+
+	void scan(Method method, Class<?> type, TypePattern<?, M> pattern) throws Exception;
+
 }
