+++ java-utils-scanner/src/main/java/net/runeduniverse/lib/utils/scanner/templates/FieldScanner.java
@@ -17,8 +17,8 @@ package net.runeduniverse.lib.utils.scanner.templates;
 
 import java.lang.reflect.Field;
 
-import net.runeduniverse.lib.utils.scanner.IFieldScanner;
 import net.runeduniverse.lib.utils.scanner.ScanOrder;
+import net.runeduniverse.lib.utils.scanner.api.IFieldScanner;
 import net.runeduniverse.lib.utils.scanner.pattern.FieldPattern;
 import net.runeduniverse.lib.utils.scanner.pattern.TypePattern;
 
@@ -55,10 +55,11 @@ public class FieldScanner<F extends FieldPattern> implements IFieldScanner<F> {
 	}
 
 	public static FieldScanner<FieldPattern> DEFAULT() {
-		return new FieldScanner<FieldPattern>(FieldScanner::createPattern);
+		return new FieldScanner<>(FieldScanner::createPattern);
 	}
 
 	public static FieldScanner<FieldPattern> DEFAULT(ScanOrder order) {
-		return new FieldScanner<FieldPattern>(FieldScanner::createPattern, order);
+		return new FieldScanner<>(FieldScanner::createPattern, order);
 	}
+
 }
